Move TitleEditor to component, save when user stops typing
